requirements.txt
beautifulsoup4==4.7.1
numpy==1.16.4
Open-Tamil==0.97
urllib3==1.24.2